refactor - split Coong env helpers into separate modules
- Add app-url, database, and supabase-keys env modules
- Remove consolidated self.ts and update imports
