+++ apps/coong/src/utils/supabase/index.ts
@@ -2,7 +2,7 @@ import {createBrowserClient} from '@supabase/ssr'
 import type {Database, SupabaseClient} from '@supabase/supabase-js'
 import {createSupabaseServer} from './supabase-server'
 import {getRequestEvent, isServer} from 'solid-js/web'
-import {getSupabaseClientKeys} from 'src/env/self'
+import {getSupabaseClientKeys} from 'src/env'
 
 export const createSupabase = (): SupabaseClient<Database> => {
   const {key, url} = getSupabaseClientKeys()
